add mysql integration and swagger docs

+++ src/services/userService.js
@@ -1,9 +1,10 @@
-const { v4: uuidv4 } = require('uuid')
 const bcrypt = require('bcryptjs')
 const jwt = require('jsonwebtoken')
 
 const { addDocument, getDocument, updateDocument, deleteDocument, searchDocument } = require('./elasticService')
 const { getCache, setCache, deleteCache, publishMessage } = require('./redisService')
+
+const User = require('../models/User.2js')
 
 const INDEX = 'users'
 const JWT_SECRET = process.env.JWT_SECRET
@@ -12,12 +13,26 @@ const JWT_SECRET = process.env.JWT_SECRET
 async function register (user) {
   const { name, email, password } = user
   const hashedPassword = await bcrypt.hash(password, 10)
-  const timestamp = new Date().toISOString()
-  const id = uuidv4()
-  const response = await addDocument(INDEX, id, { id, name, email, password: hashedPassword, created_at: timestamp, updated_at: null })
+
+  const existedUser = await User.findOne({
+    where: { email }
+  })
+  if (existedUser) {
+    return { fromCache: false, data: null }
+  }
+
+  const newUser = await User.create({
+    name,
+    email,
+    hashedPassword
+  })
+
+  const id = newUser.id
+  const timestamp = newUser.created_at
+  const response = await addDocument(INDEX, id, { name, email, password: hashedPassword, created_at: timestamp, updated_at: null })
 
   await deleteCache(`${INDEX}:${id}`)
-  await publishMessage('user_updates', JSON.stringify({ action: 'create', id, name }))
+  await publishMessage('user_updates', JSON.stringify({ action: 'register', id, name, email }))
 
   return { fromCache: false, data: response }
 }
@@ -55,18 +70,40 @@ async function getUser (id) {
 
 // Update user
 async function updateUser (id, user) {
-  const { name, email } = user
+  const { name, email, password } = user
+  const hashedPassword = await bcrypt.hash(password, 10)
   const timestamp = new Date().toISOString()
-  const response = await updateDocument(INDEX, id, { id, name, email, updated_at: timestamp })
+  const existedUser = await User.findOne({
+    where: { id }
+  })
+  if (!existedUser) {
+    return { fromCache: false, data: null }
+  }
+  await User.update({
+    name,
+    email,
+    password: hashedPassword
+  }, {
+    where: {
+      id
+    }
+  })
+
+  const response = await updateDocument(INDEX, id, { name, email, password: hashedPassword, updated_at: timestamp })
 
   await deleteCache(`${INDEX}:${id}`)
-  await publishMessage('user_updates', JSON.stringify({ action: 'update', id, name }))
+  await publishMessage('user_updates', JSON.stringify({ action: 'update', id, name, email }))
 
   return { fromCache: false, data: response }
 }
 
 // Hapus user
 async function deleteUser (id) {
+  await User.destroy({
+    where: {
+      id
+    }
+  })
   const response = await deleteDocument(INDEX, id)
 
   await deleteCache(`${INDEX}:${id}`)
